refactor: folder structure
Just moved the utils folder and the components folder to the root of the project.
