+++ components/activity/Content.tsx
