+++ components/activity/Root.tsx
