+++ components/icons/PasswordIcon.tsx
